image_utils/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
pillow==5.1.0
pika==0.12.0
numpy==1.16.2
opencv-python==3.4.1.15
tensorflow==1.13.1
healthcheck==1.3.3
dlib==19.15.0
scikit-learn==0.19.2
scipy==1.1.0
requests===2.19.1
